pyproject.toml
[project]
name = "api-extraer-info-texto"
version = "0.0.1"
dependencies = [
  "uvicorn==0.34.0", # Para ejecutar el servicio
  "fastapi==0.115.11",
  "python-dotenv",
  "langchain-aws==0.2.16",
  "langchain-community==0.3.20",
  "PyJWT"
]
requires-python = ">=3.11"
authors = [
  {name = "Andres Ortiz", email = "candresortizm@gmail.com"}
]
maintainers = [
  {name = "Andres Ortiz", email = "candresortizm@gmail.com"}
]
description = "Proyecto de una API REST que usa un LLM para extraer un resumen y entidades clave de un texto."
readme = "README.md"